Move BodySourceMap to dedicated module.
