+++ crates/ra_hir_def/src/body.rs
@@ -1,16 +1,16 @@
 //! Defines `Body`: a lowered representation of bodies of functions, statics and
 //! consts.
 mod lower;
+mod source_map;
 pub mod scope;
 
 use std::{ops::Index, sync::Arc};
 
 use hir_expand::{
-    either::Either, hygiene::Hygiene, AstId, HirFileId, InFile, MacroDefId, MacroFileKind,
+    hygiene::Hygiene, AstId, HirFileId, InFile, MacroDefId, MacroFileKind,
 };
-use ra_arena::{map::ArenaMap, Arena};
-use ra_syntax::{ast, AstNode, AstPtr};
-use rustc_hash::FxHashMap;
+use ra_arena::Arena;
+use ra_syntax::{ast, AstNode};
 
 use crate::{
     db::DefDatabase,
@@ -20,6 +20,11 @@ use crate::{
     src::HasSource,
     DefWithBodyId, HasModule, Lookup, ModuleId,
 };
+pub use source_map::{
+    ExprPtr, ExprSource, PatPtr, PatSource,
+    BodySourceMap,
+};
+pub(crate) use source_map::BodyWithSourceMap;
 
 struct Expander {
     crate_def_map: Arc<CrateDefMap>,
@@ -115,32 +120,6 @@ pub struct Body {
     pub body_expr: ExprId,
 }
 
-pub type ExprPtr = Either<AstPtr<ast::Expr>, AstPtr<ast::RecordField>>;
-pub type ExprSource = InFile<ExprPtr>;
-
-pub type PatPtr = Either<AstPtr<ast::Pat>, AstPtr<ast::SelfParam>>;
-pub type PatSource = InFile<PatPtr>;
-
-/// An item body together with the mapping from syntax nodes to HIR expression
-/// IDs. This is needed to go from e.g. a position in a file to the HIR
-/// expression containing it; but for type inference etc., we want to operate on
-/// a structure that is agnostic to the actual positions of expressions in the
-/// file, so that we don't recompute types whenever some whitespace is typed.
-///
-/// One complication here is that, due to macro expansion, a single `Body` might
-/// be spread across several files. So, for each ExprId and PatId, we record
-/// both the HirFileId and the position inside the file. However, we only store
-/// AST -> ExprId mapping for non-macro files, as it is not clear how to handle
-/// this properly for macros.
-#[derive(Default, Debug, Eq, PartialEq)]
-pub struct BodySourceMap {
-    expr_map: FxHashMap<ExprSource, ExprId>,
-    expr_map_back: ArenaMap<ExprId, ExprSource>,
-    pat_map: FxHashMap<PatSource, PatId>,
-    pat_map_back: ArenaMap<PatId, PatSource>,
-    field_map: FxHashMap<(ExprId, usize), AstPtr<ast::RecordField>>,
-}
-
 impl Body {
     pub(crate) fn body_with_source_map_query(
         db: &impl DefDatabase,
@@ -198,29 +177,5 @@ impl Index<PatId> for Body {
 
     fn index(&self, pat: PatId) -> &Pat {
         &self.pats[pat]
-    }
-}
-
-impl BodySourceMap {
-    pub fn expr_syntax(&self, expr: ExprId) -> Option<ExprSource> {
-        self.expr_map_back.get(expr).copied()
-    }
-
-    pub fn node_expr(&self, node: InFile<&ast::Expr>) -> Option<ExprId> {
-        let src = node.map(|it| Either::A(AstPtr::new(it)));
-        self.expr_map.get(&src).cloned()
-    }
-
-    pub fn pat_syntax(&self, pat: PatId) -> Option<PatSource> {
-        self.pat_map_back.get(pat).copied()
-    }
-
-    pub fn node_pat(&self, node: InFile<&ast::Pat>) -> Option<PatId> {
-        let src = node.map(|it| Either::A(AstPtr::new(it)));
-        self.pat_map.get(&src).cloned()
-    }
-
-    pub fn field_syntax(&self, expr: ExprId, field: usize) -> AstPtr<ast::RecordField> {
-        self.field_map[&(expr, field)]
     }
 }
